+++ src/zope/cachedescriptors/property.py
@@ -19,8 +19,9 @@ from functools import update_wrapper
 ncaches = 0
 
 
-class CachedProperty(object):
-    """Cached Properties.
+class _CachedProperty(object):
+    """
+    Cached property implementation class.
     """
 
     def __init__(self, func, *names):
@@ -53,6 +54,36 @@ class CachedProperty(object):
         setattr(inst, value_name, value)
 
         return value
+
+def CachedProperty(*args):
+    """
+    CachedProperties.
+
+    This is usable directly as a decorator when given names, or when not. Any of these patterns
+    will work:
+
+    * ``@CachedProperty``
+    * ``@CachedProperty()``
+    * ``@CachedProperty('n','n2')``
+    * def thing(self: ...; thing = CachedProperty(thing)
+    * def thing(self: ...; thing = CachedProperty(thing, 'n')
+
+    """
+
+    if not args:  # @CachedProperty()
+        return _CachedProperty  # A callable that produces the decorated function
+
+    arg1 = args[0]
+    names = args[1:]
+    if callable(arg1):  # @CachedProperty, *or* thing = CachedProperty(thing, ...)
+        return _CachedProperty(arg1, *names)
+
+    # @CachedProperty( 'n' )
+    # Ok, must be a list of string names. Which means we are used like a factory
+    # so we return a callable object to produce the actual decorated function
+    def factory(function):
+        return _CachedProperty(function, arg1, *names)
+    return factory
 
 
 class Lazy(object):